finalizando proyecto

+++ database/migrations/2021_04_03_042730_create_compras_table.php
@@ -18,7 +18,7 @@ class CreateComprasTable extends Migration
 
             $table->string('serial');
             $table->string('registro');
-            $table->integer('presentacion');
+            $table->decimal('presentacion');
             $table->string('color');
